+++ examples/v1/plot_ca.py
@@ -12,14 +12,17 @@ def plot_ca(ca_files, labels, ca_img=None):
     fig, ax = plt.subplots()
 
     for path, label in zip(ca_files, labels):
-        df = pd.read_csv(path, index_col=[0, 1])
+        df = pd.read_csv(path, index_col="name")
+        df["file"] = (
+            df.index.to_series().astype(str).str.rsplit("/", n=1).str[0].to_numpy()
+        )
 
         all_t, all_y = [], []
         for _, subdf in df.groupby("file"):
-            t = pd.to_timedelta(subdf["Rec. Time"]).dt.total_seconds()
+            t = pd.to_timedelta(subdf["record_time"]).dt.total_seconds()
             t = t - t.iloc[0]
             all_t.append(t.values)
-            all_y.append(subdf["Contact Angle(Average)[degree]"].values)
+            all_y.append(subdf["contact_angle"].values)
 
         t_ref = np.unique(np.concatenate(all_t))
         y_interp = np.array(