+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v1.0.0b1] - 2026-07-21
+## UNRELEASED
 
 Includes:
 
@@ -17,3 +17,4 @@ Includes:
   - Preprocessed sets of profiles
   - Mean profiles of each set of profiles
   - Process variables
+  - Contact angle measurement data